accept package layout inspect
